Implemented the find_py4j functions into the Nexus Base class
